fix: Controller, Service, Repository수정
# 예외처리 Sercice -> Controller로 이동
# ResponseDto create함수명 짧게 수정
# 임시 테스트 쿠키 데이터 추가
# 오류가 발생하지 않도록 현재 사용되지 않는 코드 주석처리

+++ src/main/java/com/sparta/schedule_project/service/CommentService.java
@@ -1,5 +1,6 @@
 package com.sparta.schedule_project.service;
 
+import com.sparta.schedule_project.common.TestCookieData;
 import com.sparta.schedule_project.dto.request.comment.CreateCommentRequestDto;
 import com.sparta.schedule_project.dto.request.comment.ModifyCommentRequestDto;
 import com.sparta.schedule_project.dto.request.comment.RemoveCommentRequestDto;
@@ -7,7 +8,9 @@ import com.sparta.schedule_project.dto.request.comment.SearchCommentRequestDto;
 import com.sparta.schedule_project.dto.response.ResponseStatusDto;
 import com.sparta.schedule_project.dto.response.comment.CommentResponseDto;
 import com.sparta.schedule_project.entity.Comment;
+import com.sparta.schedule_project.entity.Schedule;
 import com.sparta.schedule_project.exception.ResponseCode;
+import com.sparta.schedule_project.exception.ResponseException;
 import com.sparta.schedule_project.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
@@ -27,13 +30,9 @@ public class CommentService {
      * @since 2024-10-15
      */
     public ResponseStatusDto createComment(CreateCommentRequestDto requestDto) {
-        try {
-            Comment comment = CreateCommentRequestDto.to(requestDto);
-            commentRepository.save(comment);
-            return new ResponseStatusDto(ResponseCode.SUCCESS_CREATE_SCHEDULE);
-        } catch (Exception ex) {
-            return new ResponseStatusDto(ResponseCode.UNKNOWN_ERROR, ex.getMessage());
-        }
+        Comment comment = CreateCommentRequestDto.to(requestDto);
+        commentRepository.save(comment);
+        return new ResponseStatusDto(ResponseCode.SUCCESS_CREATE_SCHEDULE);
     }
 
     /**
@@ -45,13 +44,9 @@ public class CommentService {
      * @since 2024-10-15
      */
     public CommentResponseDto searchComment(SearchCommentRequestDto requestDto) {
-        try {
-            Comment comment = SearchCommentRequestDto.to(requestDto);
-            Page<Comment> comments = commentRepository.findAllByScheduleOrderByUpdateDateDesc(comment.getSchedule(), comment.getPage());
-            return CommentResponseDto.createCommentResponseDto(comments, ResponseCode.SUCCESS_SEARCH_SCHEDULE);
-        } catch (Exception ex) {
-            return CommentResponseDto.createCommentResponseDto(ResponseCode.UNKNOWN_ERROR, ex.getMessage());
-        }
+        Comment comment = SearchCommentRequestDto.to(requestDto);
+        Page<Comment> comments = commentRepository.findAllByScheduleOrderByUpdateDateDesc(comment.getSchedule(), comment.getPage());
+        return CommentResponseDto.createResponseDto(comments, ResponseCode.SUCCESS_SEARCH_SCHEDULE);
     }
 
     /**
@@ -62,19 +57,12 @@ public class CommentService {
      * @author 김현정
      * @since 2024-10-15
      */
-    public ResponseStatusDto updateComment(ModifyCommentRequestDto requestDto) {
-        try {
-            Comment updateInfo = ModifyCommentRequestDto.to(requestDto);
-            Comment schedule = commentRepository.findBySeq(updateInfo.getSeq());
-            schedule.update(updateInfo);
-            return new ResponseStatusDto(ResponseCode.SUCCESS_UPDATE_SCHEDULE);
-        }
-//        catch (ResponseException ex) {
-//            return new ResponseStatusDto(ex.getResponseCode());
-//        }
-        catch (Exception ex) {
-            return new ResponseStatusDto(ResponseCode.UNKNOWN_ERROR, ex.getMessage());
-        }
+    public ResponseStatusDto updateComment(ModifyCommentRequestDto requestDto) throws ResponseException {
+        Comment updateInfo = ModifyCommentRequestDto.to(requestDto);
+        Comment comment = commentRepository.findBySeq(updateInfo.getSeq());
+        checkAccess(comment);
+        comment.update(updateInfo);
+        return new ResponseStatusDto(ResponseCode.SUCCESS_UPDATE_SCHEDULE);
     }
 
     /**
@@ -85,19 +73,23 @@ public class CommentService {
      * @author 김현정
      * @since 2024-10-15
      */
-    public ResponseStatusDto deleteSchedule(RemoveCommentRequestDto requestDto) {
-        try {
-//            checkAccess(requestDto);
-            Comment comment = RemoveCommentRequestDto.to(requestDto);
-            commentRepository.delete(comment);
-            return new ResponseStatusDto(ResponseCode.SUCCESS_DELETE_SCHEDULE);
-        }
-//        catch (ResponseException ex)
-//        {
-//            return new ResponseStatusDto(ex.getResponseCode());
-//        }
-        catch (Exception ex) {
-            return new ResponseStatusDto(ResponseCode.UNKNOWN_ERROR, ex.getMessage());
-        }
+    public ResponseStatusDto deleteSchedule(RemoveCommentRequestDto requestDto) throws ResponseException {
+        Comment comment = RemoveCommentRequestDto.to(requestDto);
+        checkAccess(comment);
+        commentRepository.delete(comment);
+        return new ResponseStatusDto(ResponseCode.SUCCESS_DELETE_SCHEDULE);
+    }
+
+    /**
+     * 댓글 수정/삭제 요청에 대한 권한을 검사합니다.
+     *
+     * @param comment 일정 수정 권한
+     * @throws ResponseException 권한이 없는 경우 예외를 발생시킵니다.
+     * @author 김현정
+     * @since 2024-10-15
+     */
+    private void checkAccess(Comment comment) throws ResponseException {
+        if (TestCookieData.testSeq != comment.getUser().getSeq())
+            throw new ResponseException(ResponseCode.INVALID_PERMISSION);
     }
 }
